requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
idna==3.4
pydantic==2.1.1
pydantic-core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
